+++ src/NativeScreenSecurity.ts
@@ -1,8 +1,10 @@
 import { TurboModule, TurboModuleRegistry } from 'react-native';
 
+export type BlurStyleEnum = 'light' | 'dark' | 'system' | 'extraLight';
+
 export interface Spec extends TurboModule {
   setSecureWindow(enable: boolean): void;
-  setAppSwitcherBlur(enable: boolean, style: string): void;
+  setAppSwitcherBlur(enable: boolean, style: BlurStyleEnum): void;
   addListener(eventName: string): void;
   removeListeners(count: number): void;
 }